+++ NEWS.md
@@ -66,7 +66,7 @@
 * `multi_line` attribute in parse table is now integer, not boolean (#782).
 * style guide used in Addin is verified when set via R option (#789).
 * improve pkgdown author URLs (#775).
-* Upgrade touchstone infra (#799).
+* Upgrade touchstone infra (#799, #805).
 * Don't test on R 3.3. anymore as tidyverse 
   [supports only four previous releases](https://www.tidyverse.org/blog/2019/04/r-version-support/) (#804).